Former-commit-id: 20c066c44190d4a4c0c392b6130065ad76d306b7

+++ web/concrete/models/job.php
@@ -240,7 +240,7 @@ class Job extends Object {
 	}
 	
 	final function executeJob(){
-		
+		Events::fire('on_before_job_execute', $this);
 		$db = Loader::db();
 		$timestampH =date('Y-m-d g:i:s A');
 		$timestamp=date('Y-m-d H:i:s');
@@ -264,7 +264,7 @@ class Job extends Object {
 			$enum = $this->getError();
 		}
 		$rs = $db->query( "INSERT INTO JobsLog (jID, jlMessage, jlTimestamp, jlError) VALUES(?,?,?,?)", array( $this->jID, $resultMsg, $timestamp, $enum ) );
-		
+		Events::fire('on_job_execute', $this);
 		
 		return $resultMsg;
 	}
@@ -305,12 +305,13 @@ class Job extends Object {
 			$db = Loader::db();
 			$db->Execute('insert into Jobs (jName, jDescription, jDateInstalled, jNotUninstallable, jHandle, pkgID) values (?, ?, ?, ?, ?, ?)', 
 				array($j->getJobName(), $j->getJobDescription(), Loader::helper('date')->getLocalDateTime(), 0, $jHandle, $pkg->getPackageID()));
-			
+			Events::fire('on_job_install' $j);
 			return $j;
 		}
 	}
  
 	final public function install(){
+		
 		$db = Loader::db();
 		$jobExists=$db->getOne( 'SELECT count(*) FROM Jobs WHERE jHandle=?', array($this->jHandle) );
 		$vals=array($this->getJobName(),$this->getJobDescription(),  date('Y-m-d H:i:s'), $this->jNotUninstallable, $this->jHandle);
@@ -319,9 +320,14 @@ class Job extends Object {
 		}else{
 			$db->query('INSERT INTO Jobs (jName, jDescription, jDateInstalled, jNotUninstallable, jHandle) VALUES(?,?,?,?,?)',$vals);
 		}
+		Events::fire('on_job_install' $this);
 	}
  
 	final public function uninstall(){
+		$ret = Events::fire('on_job_uninstall', $this);
+		if($ret < 0) {
+			return $ret;
+		}
 		$db = Loader::db();
 		$db->query( 'DELETE FROM Jobs WHERE jHandle=?', array($this->jHandle) );
 	}
